Adjust formatting and spelling as per the DEC edition: AMAZIN, KING, LETTER.
* Update KING.BAS

Fixed all of the mistakes. In addition, I argue that the "THINK AGAIN" lines have commas, not periods, because they consistently seem to have a "stroke" towards the bottom-left (I don't know the exact terminology), whereas the periods in "SQ.", for example, don't have that.

* Update LETTER.BAS

The lines 460 and 480 do not have two spaces after the periods, it just looks like that because of the font.

* Update LETTER.BAS

Changed 420 to 425 to match original program, and removed one "!"

* Update AMAZIN.BAS

One space removed. The transcription is otherwise accurate