Update BOAT.BAS
My most difficult transcription so far. It's challenging to figure out the exact shapes of the right edges of the boat due to the printing errors; thankfully, the sample run provides proof of the shapes of two of the four possible bodies of the boat (1st and 3rd). The rest however have to be determined with educated guesses, unless a better source can be found, however, I couldn't find any.

The book doesn't show how long the top row of the 4th boat body is, only how many Es it has at minimum, so it can just be left as is.

As for the 2nd boat body, although there is no direct evidence of the shape, I argue that the best choice is to add one more E to the top row, because:
- It evens out the spacing of the edge of the boat, which makes sense from an aesthetic perspective
- With this change, the 2nd body shares the same edge shape with the 1st body, whereas as it currently is, it doesn't share the shape with any other body. This makes the art more consistent (although it isn't very consistent elsewhere).

The rest of the changes were just basic transcription error fixes.