Update CAN-AM.BAS
This one is tough to figure out because O's and zeros look identical on the font that this was printed out on, and my inability to immediately test the program.

I have reason to believe those are Os instead of 0s on lines 850 and 860, since both O9 and O1 are variables, and the numbers 1 and 9 are not typed like this anywhere else in the program as far as I know.

There are many lines where a zero could be mistaken for an O and vice versa due to the font (for example lines 980 and 990), and honestly I'm not familiar enough with neither the program not this specific dialect, so I'm just trusting that they're zeros because there doesn't seem to be a variable called just "O".

For line 1130, I'm leaning more towards the second L being faded out rather than it being a typo, due to:
- The spelling being pristine everywhere else in the program
- Switching an L for a second space bar press would be an odd typo to make
- The first L lacks the "hook", hinting at the section being faded out
But honestly, it's your call whether you think it's faded or not.

Line 1780 is a complete mystery to me, because 1E is not a variable in the program, and I'm not aware of "LE" being a keyword of any kind. Maybe just a typo, but I doubt it.

Everything else is standard fixes.