Big batch of fixes (#7)
* Update BATTLE.BAS

* Update BINGO.BAS

* Update BINGO.BAS

Aligned formatting with original program.

* Update BLKJAC.BAS

* Update BLKJAC.BAS

Align formatting with original program.

* Update BLKJAK.BAS

* Update BLKJAK.BAS

GOSUB-->GOTO on line 115

* Update BOAT.BAS

My most difficult transcription so far. It's challenging to figure out the exact shapes of the right edges of the boat due to the printing errors; thankfully, the sample run provides proof of the shapes of two of the four possible bodies of the boat (1st and 3rd). The rest however have to be determined with educated guesses, unless a better source can be found, however, I couldn't find any.

The book doesn't show how long the top row of the 4th boat body is, only how many Es it has at minimum, so it can just be left as is.

As for the 2nd boat body, although there is no direct evidence of the shape, I argue that the best choice is to add one more E to the top row, because:
- It evens out the spacing of the edge of the boat, which makes sense from an aesthetic perspective
- With this change, the 2nd body shares the same edge shape with the 1st body, whereas as it currently is, it doesn't share the shape with any other body. This makes the art more consistent (although it isn't very consistent elsewhere).

The rest of the changes were just basic transcription error fixes.

* Update BOAT.BAS

Fixed the ASCII art + a couple other minor fixes

* Update BOMBER.BAS

I don't know why there is a colon on line 370, but that's definitely a colon.

* Update BOMBER.BAS

Formatting + spelling fixes

* Update BOUNCE.BAS

* Update BOUNCE.BAS

Align formatting with original program.

* Update BOWL.BAS

* Update BOWL.BAS

Align formatting with original program.

* Update BUG.BAS

* Update BUG.BAS

Aligned formatting and strings with original program. Although I don't know whether "HARRISON JR. SR. HIGH SCHOOL" is a real school or an error.

* Update BULCOW.BAS

* Update BULEYE.BAS

* Update BULEYE.BAS

Match strings with original program

* Update BULL.BAS

* Update BULL.BAS

Aligned formatting and wording with original program.

* Update BUNNY.ASC

Restored the mistakes of the original art.

* Update CALNDR.BAS

* Update CALNDR.BAS

Align formatting with original program.

* Update CAN-AM.BAS

This one is tough to figure out because O's and zeros look identical on the font that this was printed out on, and my inability to immediately test the program.

I have reason to believe those are Os instead of 0s on lines 850 and 860, since both O9 and O1 are variables, and the numbers 1 and 9 are not typed like this anywhere else in the program as far as I know.

There are many lines where a zero could be mistaken for an O and vice versa due to the font (for example lines 980 and 990), and honestly I'm not familiar enough with neither the program not this specific dialect, so I'm just trusting that they're zeros because there doesn't seem to be a variable called just "O".

For line 1130, I'm leaning more towards the second L being faded out rather than it being a typo, due to:
- The spelling being pristine everywhere else in the program
- Switching an L for a second space bar press would be an odd typo to make
- The first L lacks the "hook", hinting at the section being faded out
But honestly, it's your call whether you think it's faded or not.

Line 1780 is a complete mystery to me, because 1E is not a variable in the program, and I'm not aware of "LE" being a keyword of any kind. Maybe just a typo, but I doubt it.

Everything else is standard fixes.

* Update CHANGE.BAS

* Update CHANGE.BAS

* Update CHECKR.BAS

* Update CHECKR.BAS

Align formatting with original program.

* Update CHIEF.BAS

* Update CHIEF.BAS

Aligned everything with original program (including missing comma on line 50)

* Update CHOMP.BAS

* Update CHOMP.BAS

Aligned strings with original program

* Update CIVILW.BAS

Single spelling error fix.

* Update CIVILW.BAS